+++ buildvenv.sh
@@ -4,3 +4,4 @@ source venv/bin/activate
 pip3 install flask
 pip3 install requests
 pip3 install bs4
+pip3 install gunicorn
